Change folder structure to remove distiinction between geo and mat
