+++ qc_openmaterial3d/checks/xom_xoma_checker/__init__.py
@@ -5,8 +5,9 @@
 # Public License, v. 2.0. If a copy of the MPL was not distributed
 # with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from . import geo_vehicle_class_data_defined as vehicle_class_data_defined
-from . import geo_human_class_data_defined as human_class_data_defined
-from . import geo_light_definition_nodes_exist as light_definition_nodes_exist
-from . import geo_emissive_light_nodes_exist as emissive_light_nodes_exist
+from . import xoma_material_textures_exist as material_textures_exist
 from . import xoma_texture_assignment_requires_mapping as texture_assignment_requires_mapping
+from . import xoma_vehicle_class_data_defined as vehicle_class_data_defined
+from . import xoma_human_class_data_defined as human_class_data_defined
+from . import xoma_light_definition_nodes_exist as light_definition_nodes_exist
+from . import xoma_emissive_light_nodes_exist as emissive_light_nodes_exist
