+++ qc_openmaterial3d/checks/xom_xomp_checker/__init__.py
@@ -4,5 +4,4 @@
 # Public License, v. 2.0. If a copy of the MPL was not distributed
 # with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from . import mat_tables_sorted_correctly as tables_sorted_correctly
-from . import mat_look_up_tables_unique_wavelengths as look_up_tables_unique_wavelengths
+from . import xomp_look_up_tables_unique_wavelengths as look_up_tables_unique_wavelengths
