+++ qc_openmaterial3d/checks/xom_xompt_checker/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright 2025, ASAM e.V.
+# This Source Code Form is subject to the terms of the Mozilla
+# Public License, v. 2.0. If a copy of the MPL was not distributed
+# with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+from . import xompt_tables_sorted_correctly as tables_sorted_correctly
