+++ tests/test_xom_xoma_checks.py
@@ -8,56 +8,106 @@
 import os
 import test_utils
 from qc_baselib import Result, IssueSeverity, StatusType
-from qc_openmaterial3d.checks import xom_geo_checker
+from qc_openmaterial3d.checks import xom_xoma_checker
 
 
-def test_texture_assignment_requires_mapping_positive(
-    monkeypatch,
-) -> None:
+def test_material_textures_exist_positive(monkeypatch) -> None:
+    base_path = "tests/data/material_textures_exist/"
+    target_file_name = "material_textures_exist.positive.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_xoma_checker.material_textures_exist.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.material_textures_exist")) == 0
+
+    test_utils.cleanup_files()
+
+
+def test_material_textures_exist_negative(monkeypatch) -> None:
+    base_path = "tests/data/material_textures_exist/"
+    target_file_name = "material_textures_exist.negative.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_xoma_checker.material_textures_exist.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.material_textures_exist")) == 1
+
+    test_utils.cleanup_files()
+
+
+def test_material_textures_exist_no_assignment(monkeypatch) -> None:
+    base_path = "tests/data/material_textures_exist/"
+    target_file_name = "material_textures_exist.no_assignment.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_xoma_checker.material_textures_exist.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.material_textures_exist")) == 0
+
+    test_utils.cleanup_files()
+
+
+def test_texture_assignment_requires_mapping_positive(monkeypatch) -> None:
     base_path = "tests/data/texture_assignment_requires_mapping/"
     target_file_name = "texture_assignment_requires_mapping.positive.xoma"
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.texture_assignment_requires_mapping.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.texture_assignment_requires_mapping.CHECKER_ID)
         == StatusType.COMPLETED
     )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.texture_assignment_requires_mapping")) == 0
 
-    assert (
-        len(result.get_issues_by_rule_uid("asam.net:xomgeo:1.0.0:xoma.texture_assignment_requires_mapping")) == 0
-    )
     test_utils.cleanup_files()
 
 
-def test_texture_assignment_requires_mapping_negative(
-    monkeypatch,
-) -> None:
+def test_texture_assignment_requires_mapping_negative(monkeypatch) -> None:
     base_path = "tests/data/texture_assignment_requires_mapping/"
     target_file_name = "texture_assignment_requires_mapping.negative.xoma"
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.texture_assignment_requires_mapping.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.texture_assignment_requires_mapping.CHECKER_ID)
         == StatusType.COMPLETED
     )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.texture_assignment_requires_mapping")) == 1
 
-    assert (
-        len(result.get_issues_by_rule_uid("asam.net:xomgeo:1.0.0:xoma.texture_assignment_requires_mapping")) == 1
-    )
     test_utils.cleanup_files()
 
 
@@ -67,25 +117,16 @@ def test_vehicle_class_data_defined_positive(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.vehicle_class_data_defined.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.vehicle_class_data_defined.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    assert (
-        len(
-            result.get_issues_by_rule_uid(
-                "asam.net:xomgeo:1.0.0:xoma.vehicle_class_data_defined"
-            )
-        )
-        == 0
-    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.vehicle_class_data_defined")) == 0
 
     test_utils.cleanup_files()
 
@@ -96,75 +137,16 @@ def test_vehicle_class_data_defined_negative(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.vehicle_class_data_defined.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.vehicle_class_data_defined.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    issues = result.get_issues_by_rule_uid(
-        "asam.net:xomgeo:1.0.0:xoma.vehicle_class_data_defined"
-    )
-    assert len(issues) == 1
-    assert issues[0].level == IssueSeverity.ERROR
-
-    test_utils.cleanup_files()
-
-
-def test_human_class_data_defined_positive(monkeypatch) -> None:
-    base_path = "tests/data/human_class_data_defined/"
-    target_file_name = "human_class_data_defined.positive.xoma"
-    target_file_path = os.path.join(base_path, target_file_name)
-
-    test_utils.create_test_config(target_file_path)
-
-    test_utils.launch_main(monkeypatch)
-
-    result = Result()
-    result.load_from_file(test_utils.REPORT_FILE_PATH)
-
-    assert (
-        result.get_checker_status(xom_geo_checker.human_class_data_defined.CHECKER_ID)
-        == StatusType.COMPLETED
-    )
-
-    assert (
-        len(
-            result.get_issues_by_rule_uid(
-                "asam.net:xomgeo:1.0.0:xoma.human_class_data_defined"
-            )
-        )
-        == 0
-    )
-
-    test_utils.cleanup_files()
-
-
-def test_human_class_data_defined_negative(monkeypatch) -> None:
-    base_path = "tests/data/human_class_data_defined/"
-    target_file_name = "human_class_data_defined.negative.xoma"
-    target_file_path = os.path.join(base_path, target_file_name)
-
-    test_utils.create_test_config(target_file_path)
-
-    test_utils.launch_main(monkeypatch)
-
-    result = Result()
-    result.load_from_file(test_utils.REPORT_FILE_PATH)
-
-    assert (
-        result.get_checker_status(xom_geo_checker.human_class_data_defined.CHECKER_ID)
-        == StatusType.COMPLETED
-    )
-
-    issues = result.get_issues_by_rule_uid(
-        "asam.net:xomgeo:1.0.0:xoma.human_class_data_defined"
-    )
+    issues = result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.vehicle_class_data_defined")
     assert len(issues) == 1
     assert issues[0].level == IssueSeverity.ERROR
 
@@ -177,25 +159,58 @@ def test_vehicle_class_data_defined_non_vehicle(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.vehicle_class_data_defined.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.vehicle_class_data_defined.CHECKER_ID)
         == StatusType.COMPLETED
     )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.vehicle_class_data_defined")) == 0
+
+    test_utils.cleanup_files()
+
+
+def test_human_class_data_defined_positive(monkeypatch) -> None:
+    base_path = "tests/data/human_class_data_defined/"
+    target_file_name = "human_class_data_defined.positive.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        len(
-            result.get_issues_by_rule_uid(
-                "asam.net:xomgeo:1.0.0:xoma.vehicle_class_data_defined"
-            )
-        )
-        == 0
+        result.get_checker_status(xom_xoma_checker.human_class_data_defined.CHECKER_ID)
+        == StatusType.COMPLETED
     )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.human_class_data_defined")) == 0
+
+    test_utils.cleanup_files()
+
+
+def test_human_class_data_defined_negative(monkeypatch) -> None:
+    base_path = "tests/data/human_class_data_defined/"
+    target_file_name = "human_class_data_defined.negative.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_xoma_checker.human_class_data_defined.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+    issues = result.get_issues_by_rule_uid("asam.net:xom:1.0.0:xoma.human_class_data_defined")
+    assert len(issues) == 1
+    assert issues[0].level == IssueSeverity.ERROR
 
     test_utils.cleanup_files()
 
@@ -206,25 +221,16 @@ def test_light_definition_nodes_exist_positive(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.light_definition_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.light_definition_nodes_exist.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    assert (
-        len(
-            result.get_issues_by_rule_uid(
-                "asam.net:xomgeo:1.1.0:xoma.light_definition_nodes_exist"
-            )
-        )
-        == 0
-    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.1.0:xoma.light_definition_nodes_exist")) == 0
 
     test_utils.cleanup_files()
 
@@ -235,20 +241,16 @@ def test_light_definition_nodes_exist_negative(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.light_definition_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.light_definition_nodes_exist.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    issues = result.get_issues_by_rule_uid(
-        "asam.net:xomgeo:1.1.0:xoma.light_definition_nodes_exist"
-    )
+    issues = result.get_issues_by_rule_uid("asam.net:xom:1.1.0:xoma.light_definition_nodes_exist")
     assert len(issues) == 1
     assert issues[0].level == IssueSeverity.ERROR
 
@@ -261,14 +263,13 @@ def test_light_definition_nodes_exist_skipped_old_version(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.light_definition_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.light_definition_nodes_exist.CHECKER_ID)
         == StatusType.SKIPPED
     )
 
@@ -281,25 +282,16 @@ def test_emissive_light_nodes_exist_positive(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.emissive_light_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.emissive_light_nodes_exist.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    assert (
-        len(
-            result.get_issues_by_rule_uid(
-                "asam.net:xomgeo:1.1.0:xoma.emissive_light_nodes_exist"
-            )
-        )
-        == 0
-    )
+    assert len(result.get_issues_by_rule_uid("asam.net:xom:1.1.0:xoma.emissive_light_nodes_exist")) == 0
 
     test_utils.cleanup_files()
 
@@ -310,20 +302,16 @@ def test_emissive_light_nodes_exist_negative(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.emissive_light_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.emissive_light_nodes_exist.CHECKER_ID)
         == StatusType.COMPLETED
     )
-
-    issues = result.get_issues_by_rule_uid(
-        "asam.net:xomgeo:1.1.0:xoma.emissive_light_nodes_exist"
-    )
+    issues = result.get_issues_by_rule_uid("asam.net:xom:1.1.0:xoma.emissive_light_nodes_exist")
     assert len(issues) == 1
     assert issues[0].level == IssueSeverity.ERROR
 
@@ -336,14 +324,13 @@ def test_emissive_light_nodes_exist_skipped_old_version(monkeypatch) -> None:
     target_file_path = os.path.join(base_path, target_file_name)
 
     test_utils.create_test_config(target_file_path)
-
     test_utils.launch_main(monkeypatch)
 
     result = Result()
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-        result.get_checker_status(xom_geo_checker.emissive_light_nodes_exist.CHECKER_ID)
+        result.get_checker_status(xom_xoma_checker.emissive_light_nodes_exist.CHECKER_ID)
         == StatusType.SKIPPED
     )
